requirements.txt
Django==1.9.7
djangorestframework==3.3.3
phonenumbers==7.4.4
PyJWT==1.4.0
wheel==0.24.0
